chore(hs): Filesystem Cleanup (#168)
* removed a lot of code already

* more cleanup

* removed migration

* fmt
