+++ pubky-homeserver/src/persistence/files/mod.rs
@@ -3,12 +3,10 @@ mod file_io_error;
 mod file_metadata;
 mod file_service;
 mod file_stream_type;
-mod lmdb_to_opendal_migrator;
 mod opendal_service;
 
 pub use file_io_error::{FileIoError, WriteStreamError};
 pub(crate) use file_metadata::{FileMetadata, FileMetadataBuilder};
 pub use file_service::FileService;
 pub use file_stream_type::FileStream;
-pub use lmdb_to_opendal_migrator::LmDbToOpendalMigrator;
 pub use opendal_service::OpendalService;
