--- pubky-homeserver/src/persistence/lmdb/tables/files/mod.rs
@@ -1,7 +1,0 @@
-mod blobs;
-mod entries;
-mod in_db_file;
-
-pub use blobs::{BlobsTable, BLOBS_TABLE};
-pub use entries::{EntriesTable, Entry, EntryHash, FileLocation, ENTRIES_TABLE};
-pub use in_db_file::*;
